chore: comprehensive review fixes — security, performance, CI/CD, quality (v0.32.3)
Security:
- Require AXON_WEB_ALLOW_INSECURE_DEV=true for debug auth bypass (implicit on missing token before)
- Add startup tracing::warn! when insecure dev mode is active; add 2 tests
- Shell PTY: structured audit log on session start/end (session_id, duration_ms)
- clear_collection_mode_cache() called after axon migrate for both src and dst collections

Performance:
- Graph worker Stage 1 Neo4j writes parallelized with tokio::join!() (~30-50% speedup)
  Bonus: fixes latent bug where write_entity_relationships ran before write_entities
- Embed worker: share Redis connection at startup via Arc<Mutex> (no per-job TCP handshake)
- Embed worker: progress Postgres UPDATEs debounced to 500ms (was one UPDATE per document)
- TEI: env::var() per batch replaced with LazyLock (eliminates per-call global process lock)
- Qdrant: exponential backoff + jitter on all 4 retry sites (prevents thundering herd)

CI/CD:
- Standardize all CI jobs to rust-toolchain.toml pin 1.94.0 (drop mixed 1.93.1 usages)
- Add weekly scheduled run Mon 03:00 UTC for AMQP infra integration tests
- Switch cargo-audit/cargo-deny to taiki-e/install-action prebuilt binaries (cached)
- Renovate regexManagers for YT_DLP, S6_OVERLAY, CLAUDE_ACP, CODEX_ACP Dockerfile ARGs
- Add dependencyDashboardApproval for Rust major/minor updates
- Protect services.env with pre-commit env guard (*.env pattern in check_env_staged.sh)
- Remove criterion dev-dep (no benches/ directory)

Code quality:
- Deprecate open_amqp_channel() + narrow to pub(crate) (footgun: connection dropped immediately)
- Add #[must_use] to 26 public service entry-point functions
- Replace manual Display+Error impls on PayloadParseError with thiserror derive
- Standardize named exports in terminal-emulator.tsx and neural-canvas-core.tsx
- Fix f64::EPSILON as f32 → f32::EPSILON in graph similarity test
- Fix snippet.rs unwrap() → let-else guard on iterator next()

Testing:
- 5 SQL safety tests: JobTable/JobStatus values validated as alphanumeric+underscore
- 2 collection mode cache tests (clear_cache_allows_mode_re_detection, noop_for_absent_entry)
- 2 auth bypass tests; CI schedule now gates infra tests on schedule or manual dispatch

Documentation:
- spider-feature-flags.md: remove glob (excluded), add hedge, update spider_agent 2.45→2.46
- .env.example: AXON_COLLECTION=axon → cortex (matches code default and all other docs)
- docs/SECURITY.md: document AXON_WEB_ALLOW_INSECURE_DEV loopback bypass in Residual Risks
- docs/auth/API-TOKEN.md: add AXON_WEB_BROWSER_API_TOKEN; fix NEXT_PUBLIC guidance
- docs/ACP.md: label SESSION_TTL/MAX_REPLAY_BUFFER as hardcoded (not env-configurable)
- CLAUDE.md: fix AMQP reconnect file path; clarify run_*_native() as API surface removal

+++ crates/jobs/graph/similarity.rs
@@ -218,6 +218,6 @@ mod tests {
             .iter()
             .find(|edge| edge.target_url == "https://b.com")
             .unwrap();
-        assert!((b.score - 0.91).abs() < f64::EPSILON as f32);
+        assert!((b.score - 0.91).abs() < f32::EPSILON);
     }
 }
